INTEGER type in sqlite can also be a long apparently.

+++ src/main/java/no/kvikshaug/gh/modules/Timer.java
@@ -37,7 +37,7 @@ public class Timer implements TriggerListener {
     		for(Object[] row : rows) {
     			int id = (Integer) row[0];
     			String nick = (String) row[1];
-    			int time = (Integer) row[2];
+    			long time = (Long) row[2];
     			String message = (String) row[3];
 
     			//The timer expired when the bot was down
